@@ -7,8 +7,8 @@ import { Project } from '../entity/project';
 import { ProjectsService } from '../services/projects.service';
 import { File } from '../entity/file';
 import { ProgressesService } from '../services/progresses.service';
+import { ExecutionConfig } from '../entity/execution-config';
 //import { SSL_OP_SSLREF2_REUSE_CERT_TYPE_BUG } from 'constants';
-
 @Component({
   selector: 'app-widget-code',
   templateUrl: './widget-code.component.html',
@@ -16,12 +16,13 @@ import { ProgressesService } from '../services/progresses.service';
 })
 
 
-export class WidgetCodeComponent extends WidgetComponent implements OnInit {  
+export class WidgetCodeComponent extends WidgetComponent implements OnInit {
 
-  output = 'Output window';
+  output: String = "";
   relativeProject: Project;
-  projectDBId: String;
+  progressProjectID: String;
   files: File[];
+  pullerTimer: NodeJS.Timer;
 
   constructor(private projectService: ProjectsService, private progressesService: ProgressesService) {
     super();
@@ -30,10 +31,10 @@ export class WidgetCodeComponent extends WidgetComponent implements OnInit {
   ngOnInit() {
     this.projectService.getProject(this.widget.projectID).subscribe(pr => {
       this.relativeProject = pr;
-      this.projectDBId = this.widget.projectID.replace("/api/projects/", "");
+      this.progressProjectID = "/api/progresses/myself/projects/" + this.widget.projectID.replace("/api/projects/", "");
       this.files = pr.files;
       this.files.forEach(f => {
-        this.progressesService.getFileOfProgressProjects("/api/progresses/myself/projects/" + this.projectDBId, f.filename)
+        this.progressesService.getFileOfProgressProjects(this.progressProjectID, f.filename)
         .subscribe(f2 => {
           this.files[this.files.findIndex(f3 => f3.filename === f.filename)].content = f2;
         });
@@ -41,17 +42,44 @@ export class WidgetCodeComponent extends WidgetComponent implements OnInit {
     });
   }
 
-  DoSomething(value: string): void{
-    console.log(value);
-    this.output=value;
-
-    //Richiesta POST verso BE
+  BuildAndRun(value: string): void {
+    //Put of the file
+    this.output = "";
+    if (this.pullerTimer) {
+      clearInterval(this.pullerTimer);
+    }
+    this.progressesService.putFileOfProgressProjects(this.progressProjectID,
+      this.getExecutionConfig("RUN").mainFile, value).subscribe(() => {
+        this.progressesService.buildProject(this.progressProjectID, "RUN").subscribe(res => {
+          this.output = this.output.concat(res.compilationOutput.toString());
+          this.progressesService.executeProject(this.progressProjectID, "RUN").subscribe(() => {
+            console.log("Executing...");
+            this.pullerTimer = setInterval(() => {
+              this.progressesService.pullStdout(this.progressProjectID).subscribe(cont => {
+                this.output = this.output.concat(cont.toString());
+                console.log("Pulled stdout");
+              });
+              this.progressesService.pullStderr(this.progressProjectID).subscribe(cont => {
+                this.output = this.output.concat(cont.toString());
+                console.log("Pulled stderr");
+              });
+            }, 500);
+          });
+        });
+      });
   }
 
   getFile(): String {
+    if (!this.files) {
+      return "";
+    }
     return this.files[this.files.findIndex(f3 => f3.filename === this.relativeProject.executionConfigs.find(
       ec => ec.name === "RUN"
     ).mainFile)].content;
+  }
+
+  getExecutionConfig(name: String): ExecutionConfig {
+    return this.relativeProject.executionConfigs.find(ec => ec.name === name);
   }
 }
 
